rag_app/requirements.txt
Flask
pypdf
sentence-transformers
faiss-cpu
requests
